requirements.txt
# Deep Learning
torch==2.0.1
torchvision==0.15.2

# NLP
transformers==4.30.2

# Machine Learning
scikit-learn==1.3.0

# Data Processing
pandas==2.0.3
numpy==1.24.3

# Image Processing
pillow==10.0.0
qrcode==7.4.2

# Utilities
requests==2.31.0
tqdm==4.65.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Jupyter (optional, for notebooks)
jupyter==1.0.0
ipykernel==6.25.0

# Kaggle API
kaggle==1.5.16
